Modify the index file
