+++ index.js
@@ -1,0 +1,7 @@
+/* eslint-disable no-unused-vars */
+// eslint-disable-next-line no-unused-vars
+import * as main from './modules/script.js';
+import * as inter from './modules/medium.js';
+import * as sections from './modules/sectionFolder.js';
+import myLibrary from './modules/medium.js';
+import displayTime from './modules/timeDisplay.js';
