+++ modules/script.js
@@ -3,7 +3,7 @@ const niceTitle = document.getElementById('title');
 const theAuthor = document.getElementById('author');
 const addButton = document.querySelector('.add-btn');
 
-class Library {
+export class Library {
   constructor() {
     this.books = [
       {
@@ -63,56 +63,6 @@ class Library {
   }
 }
 
-const myLibrary = new Library();
-addButton.addEventListener('click', (event) => {
-  event.preventDefault();
-  myLibrary.addBook(niceTitle.value, theAuthor.value);
-});
-
-document.addEventListener('DOMContentLoaded', () => {
-  myLibrary.getLocalStorage();
-});
-
-// Final-milestone js-code
-
-const section1 = document.querySelector('.section1');
-const section2 = document.querySelector('.section2');
-const section3 = document.querySelector('.section3');
-const navList = document.getElementById('book-list');
-const navAdd = document.getElementById('add-book');
-const navContact = document.getElementById('contact');
-
-const showBooks = () => {
-  section1.classList.remove('hide');
-  section2.classList.add('hide');
-  section3.classList.add('hide');
-  navList.classList.add('current');
-  navAdd.classList.remove('current');
-  navContact.classList.remove('current');
+export {
+  addButton, bookList, niceTitle, theAuthor,
 };
-
-const showAddForm = () => {
-  section2.classList.remove('hide');
-  section1.classList.add('hide');
-  section3.classList.add('hide');
-  navList.classList.remove('current');
-  navAdd.classList.add('current');
-  navContact.classList.remove('current');
-};
-const showContactInfo = () => {
-  section3.classList.remove('hide');
-  section1.classList.add('hide');
-  section2.classList.add('hide');
-  navContact.classList.add('current');
-  navAdd.classList.remove('current');
-  navList.classList.remove('current');
-};
-navContact.addEventListener('click', showContactInfo);
-navAdd.addEventListener('click', showAddForm);
-navList.addEventListener('click', showBooks);
-
-const displayTime = () => {
-  document.getElementById('date-area').innerHTML = new Date().toLocaleString();
-  setTimeout(displayTime, 1000);
-};
-displayTime();
